backend/requirements.txt
# Web framework
fastapi
uvicorn

# AI Services
replicate       # for AI model integration
openai          # backup AI service

# Data parsing
pypdf2          # simple PDF to text
python-multipart  # needed for file uploads in FastAPI

# Database
convex          # Convex database client

# Pydantic (data validation, already included in FastAPI but explicit is good)
pydantic

# Cloud Storage
boto3           # AWS S3 integration

# Environment and utilities
python-dotenv   # environment variables
tiktoken        # for token counting
